requirements.base.txt
PyGObject
PyNaCl~=1.4.0
PyYAML>=5.3.1
SQLAlchemy~=1.4.23
alembic~=1.7.1.dev0
click>=7.1
firehose~=0.5
humanize
marshmallow~=3.13.0
psycopg2>=2.7.7
python-apt>=2.2.1
python-debian~=0.1.42
pyzmq~=22.3.0
requests~=2.25.1
systemd-python~=234.0
toml~=0.10.2
tornado~=6.1
